add initial data

+++ src/main/java/ch/zli/m223/model/BookingTyp.java
@@ -1,10 +1,7 @@
 package ch.zli.m223.model;
 
 public interface BookingTyp {
-
     Long getId();
-
     String getBookingType();
-
-    Double getPrice();
+    Float getPrice();
 }
